move some Visibility stuff to bevy_camera::visibility
