+++ crates/bevy_pbr/src/lib.rs
@@ -208,10 +208,7 @@ impl Plugin for PbrPlugin {
             .register_type::<AmbientLight>()
             .register_type::<CascadeShadowConfig>()
             .register_type::<Cascades>()
-            .register_type::<CascadesVisibleEntities>()
-            .register_type::<VisibleMeshEntities>()
             .register_type::<ClusterConfig>()
-            .register_type::<CubemapVisibleEntities>()
             .register_type::<DirectionalLight>()
             .register_type::<DirectionalLightShadowMap>()
             .register_type::<NotShadowCaster>()
